cleanup MainActivity.kt, MediaInfoFragment.kt
